Move libkvconfig pkg under libkubevirt
KV configuration will still have its own
package since there are a lot of configuration
specific functions that can fit here.

Not merging it with the whole libkubevirt
helps to separate context and reduce cyclic dependency
