+++ tests/libkubevirt/config/netbinding.go
@@ -17,7 +17,7 @@
  *
  */
 
-package libkvconfig
+package config
 
 import (
 	v1 "kubevirt.io/api/core/v1"
