+++ tests/network/BUILD.bazel
@@ -53,7 +53,7 @@ go_library(
         "//tests/framework/checks:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
-        "//tests/libkvconfig:go_default_library",
+        "//tests/libkubevirt/config:go_default_library",
         "//tests/libmigration:go_default_library",
         "//tests/libnet:go_default_library",
         "//tests/libnet/cloudinit:go_default_library",
