+++ tests/storage/BUILD.bazel
@@ -62,7 +62,7 @@ go_library(
         "//tests/libdv:go_default_library",
         "//tests/libinstancetype:go_default_library",
         "//tests/libkubevirt:go_default_library",
-        "//tests/libkvconfig:go_default_library",
+        "//tests/libkubevirt/config:go_default_library",
         "//tests/libmigration:go_default_library",
         "//tests/libnet:go_default_library",
         "//tests/libnode:go_default_library",
